doc: 增加 FooterCounter 组件提高性能
