+++ src/BootstrapBlazor.Server/Components/Components/FooterCounter.razor.css
@@ -1,0 +1,7 @@
+﻿.footer-online i {
+    color: var(--bs-success)
+}
+
+.footer-online a {
+    color: #ddd;
+}
