+++ src/BootstrapBlazor.Server/Components/Layout/HomeLayout.razor.cs
@@ -18,17 +18,7 @@ public partial class HomeLayout
     [NotNull]
     private string? OS { get; set; }
 
-    private string? Runtime { get; set; }
-
     private string SelectedCulture { get; set; } = CultureInfo.CurrentUICulture.Name;
-
-    private CancellationTokenSource DisposeTokenSource { get; } = new();
-
-    [Inject]
-    [NotNull]
-    private IOptions<BootstrapBlazorOptions>? BootstrapBlazorOptions { get; set; }
-
-    private ConnectionHubOptions _options = default!;
 
     /// <summary>
     /// <inheritdoc />
@@ -51,45 +41,6 @@ public partial class HomeLayout
         {
             OS = "Unknown";
         }
-
-        _options = BootstrapBlazorOptions.Value.ConnectionHubOptions;
-        UpdateRuntime();
-    }
-
-    /// <summary>
-    /// <inheritdoc />
-    /// </summary>
-    /// <param name="firstRender"></param>
-    protected override void OnAfterRender(bool firstRender)
-    {
-        if (firstRender)
-        {
-            _ = Task.Run(async () =>
-            {
-                while (!DisposeTokenSource.IsCancellationRequested)
-                {
-                    try
-                    {
-                        await Task.Delay(1000, DisposeTokenSource.Token);
-                    }
-                    catch (TaskCanceledException)
-                    {
-
-                    }
-                    if (!DisposeTokenSource.IsCancellationRequested)
-                    {
-                        UpdateRuntime();
-                        await InvokeAsync(StateHasChanged);
-                    }
-                }
-            });
-        }
-    }
-
-    private void UpdateRuntime()
-    {
-        var ts = DateTimeOffset.Now - Cache.GetStartTime();
-        Runtime = ts.ToString("dd\\.hh\\:mm\\:ss");
     }
 
     private Task SetLang(string cultureName)
@@ -105,23 +56,5 @@ public partial class HomeLayout
         }
 
         return Task.CompletedTask;
-    }
-
-    private void Dispose(bool disposing)
-    {
-        if (disposing)
-        {
-            DisposeTokenSource.Cancel();
-            DisposeTokenSource.Dispose();
-        }
-    }
-
-    /// <summary>
-    /// <inheritdoc/>
-    /// </summary>
-    public void Dispose()
-    {
-        Dispose(true);
-        GC.SuppressFinalize(this);
     }
 }
